+++ CHANGES.md
@@ -1,5 +1,6 @@
 ## Changes (unreleased)
 
+- Fail with an explicit error message when --repo-dir is missing (#152, @gpetiot)
 - Remove not-so-useful Unicode characters from output for a11y accessibility (#151, @shindere)
 - Add missing linebreaks in the project list of activity report (#148, @gpetiot)
 - Use ocamlformat 0.21.0